powerpc/pseries: Request F/W for power of two MSI-Xs only
Current logic in requesting firmware for MSI-X interrupts
is confusing. On the one hand it is known the firmware
refuses any non power of two allocations. On the other
hand such attempts still could be taken, but surprisingly
- after the attempt to request power of two interrupts
failed, not before.

This update sticks to the idea the firmware can only
allocate power of two MSI-X interrupts and makes the logic
bit more straight.

Signed-off-by: Alexander Gordeev <agordeev@redhat.com>