percpu_ida: add an API to return free tags
add an API to return free tags, blk-mq-tag will use it.
Note, this just returns a snapshot of free tags number. blk-mq-tag has two
usages of it. One is for info output for diagnosis. The other is to quickly
check if there are free tags for request dispatch checking. Neither requires
very precise.

Signed-off-by: Shaohua Li <shli@fusionio.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>