block: use a separate plug list for blk-mq requests
blk_flush_plug_list became a bit of a mess with the introduction of
blk-mq, so I started looking into separating the blk-mq handling from
it.  Turns out that by doing this we can streamline the blk-mq
submission path a lot.

If we branch out to a blk-mq specific code path early we can do the list
sort based on the hw ctx instead of the queue and thus avoid the later
improvised loop to sort again.  In addition we can also remove the hw
irq disabling in the submission path entirely and collapse a couple of
functions in blk-mq.c, all at the cost of an additional list_head in
struct blk_plug which can go away again as soon as we remove old-school
request_fn based drivers.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Jens Axboe <axboe@kernel.dk>