dmaengine: avalon: Intel Avalon-MM DMA Interface for PCIe test
This is a sample implementation of a driver using "avalon-dma" to
perform data transfers between target device memory and system memory:

    +----------+    +----------+            +----------+
    |   RAM    |<-->|  Avalon  |<---PCIe--->|   Host   |
    +----------+    +----------+            +----------+

The target device is expected to use only Avalon-MM DMA Interface for
PCIe to initiate DMA transactions - without custom hardware specifics
to make such transfers possible.

The contents of DMAed data is not analyzed by this driver in any way.
It could be both constant or volatile and is totally dependent on the
target device implementation. Consequently, it is up for the users
to treat received or provide transmitted data.

CC: dmaengine@vger.kernel.org

Signed-off-by: Alexander Gordeev <a.gordeev.box@gmail.com>